// src/types/penguin.ts
export interface Penguin {
  id: number;
  name: string;
  scientificName: string;
  description: string;
  habitat: string;
  imageUrl: string;
  height: string;
  weight: string;
  population: string;
}

export interface PenguinFact {
  id: number;
  title: string;
  description: string;
  icon: string;
}